NeuralStyle-Transfer-App/requirements.txt
flask==2.0.1
torch==2.0.1
torchvision==0.15.2
pillow==9.5.0
numpy==1.24.3
opencv-python==4.7.0.72
requests==2.31.0
tqdm==4.65.0
python-dotenv==1.0.0
gunicorn==20.1.0
flask-cors==4.0.0
werkzeug==2.2.3 